+++ Core/Games/KerbalSpaceProgram.cs
@@ -299,6 +299,8 @@ namespace CKAN.Games.KerbalSpaceProgram
 
         public Uri MetadataBugtrackerURL => new Uri("https://github.com/KSP-CKAN/NetKAN/issues/new/choose");
 
+        public Uri DiscordURL => new Uri("https://discord.gg/65hp7G7");
+
         public Uri ModSupportURL => new Uri("https://forum.kerbalspaceprogram.com/forum/70-ksp1-technical-support-pc-modded-installs/");
 
         private static string Missions(GameInstance inst)