+++ Core/Registry/Registry.cs
@@ -637,13 +637,9 @@ namespace CKAN
         /// <see cref="IRegistryQuerier.CompatibleModules"/>
         /// </summary>
         public IEnumerable<CkanModule> CompatibleModules(StabilityToleranceConfig stabilityTolerance,
-                                                         GameVersionCriteria?     crit)
+                                                         GameVersionCriteria      crit)
             // Set up our compatibility partition
-            => crit != null ? SetCompatibleVersion(stabilityTolerance, crit).LatestCompatible
-                            : repoDataMgr?.GetAllAvailableModules(Repositories.Values)
-                                          .Select(am => am.Latest(stabilityTolerance))
-                                          .OfType<CkanModule>()
-                                         ?? Enumerable.Empty<CkanModule>();
+            => SetCompatibleVersion(stabilityTolerance, crit).LatestCompatible;
 
         /// <summary>
         /// <see cref="IRegistryQuerier.IncompatibleModules"/>
@@ -692,8 +688,7 @@ namespace CKAN
             => getAvail(identifier).Select(am => am.Latest(stabilityTolerance, gameVersion, relationshipDescriptor,
                                                            installed, toInstall))
                                    .OfType<CkanModule>()
-                                   .OrderByDescending(m => m.version)
-                                   .FirstOrDefault();
+                                   .MaxBy(m => m.version);
 
         /// <summary>
         /// Find modules with a given identifier
@@ -834,7 +829,9 @@ namespace CKAN
                     ? provs.Select(am => am.Latest(stabilityTolerance, gameVersion, relationship, installed, toInstall))
                            .OfType<CkanModule>()
                            .Where(m => m.ProvidesList.Contains(identifier))
-                           .Distinct()
+                           .GroupBy(m => m.identifier)
+                           .Select(grp => grp.MaxBy(m => m.version))
+                           .OfType<CkanModule>()
                            // Put the most popular one on top
                            .OrderByDescending(m => repoDataMgr?.GetDownloadCount(Repositories.Values, m.identifier)
                                                               ?? 0)