+++ GUI/Controls/DownloadStatistics/DownloadStatisticsPieChart.cs
@@ -19,13 +19,6 @@ namespace CKAN.GUI
     {
         public DownloadStatisticsPieChart() : base()
         {
-            Model = new PlotModel()
-            {
-                Title               = Properties.Resources.DownloadStatisticsPieChartTitle,
-                TitleColor          = SystemColors.ControlText.ToOxyColor(),
-                TextColor           = SystemColors.ControlText.ToOxyColor(),
-                PlotAreaBorderColor = SystemColors.Control.ToOxyColor(),
-            };
             Controller = new PlotController();
             // Get rid of the weird click-toooltip thing
             Controller.UnbindMouseDown(OxyMouseButton.Left);
@@ -33,6 +26,13 @@ namespace CKAN.GUI
 
         public void SetData(IReadOnlyDictionary<string, long> bytesPerHost)
         {
+            Model = new PlotModel()
+            {
+                Title               = Properties.Resources.DownloadStatisticsPieChartTitle,
+                TitleColor          = SystemColors.ControlText.ToOxyColor(),
+                TextColor           = SystemColors.ControlText.ToOxyColor(),
+                PlotAreaBorderColor = SystemColors.Control.ToOxyColor(),
+            };
             var totalSize = bytesPerHost.Values.Sum();
             var numLabels = bytesPerHost.Values.Count(size => size >= totalSize * 5 / 100) + 1;
             Model.Series.Clear();