+++ GUI/Controls/ModInfo/Contents.Designer.cs
@@ -61,7 +61,7 @@ namespace CKAN.GUI
             //
             this.NotCachedLabel.Dock = System.Windows.Forms.DockStyle.Top;
             this.NotCachedLabel.Location = new System.Drawing.Point(3, 3);
-            this.NotCachedLabel.Padding = new System.Windows.Forms.Padding(0, 6, 0, 6);
+            this.NotCachedLabel.Padding = new System.Windows.Forms.Padding(0, 0, 0, 6);
             this.NotCachedLabel.Name = "NotCachedLabel";
             this.NotCachedLabel.Size = new System.Drawing.Size(494, 30);
             this.NotCachedLabel.TabIndex = 0;