+++ GUI/Main/Main.Designer.cs
@@ -58,9 +58,11 @@ namespace CKAN.GUI
             this.GameCommandlineToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.compatibleGameVersionsToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.helpToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.gameDiscordToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.modSupportToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+            this.HelpToolStripSeparator = new System.Windows.Forms.ToolStripSeparator();
             this.userGuideToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.discordToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-            this.modSupportToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.reportClientIssueToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.reportMetadataIssueToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
             this.aboutToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
@@ -352,15 +354,31 @@ namespace CKAN.GUI
             // helpToolStripMenuItem
             //
             this.helpToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+            this.gameDiscordToolStripMenuItem,
+            this.modSupportToolStripMenuItem,
+            this.HelpToolStripSeparator,
             this.userGuideToolStripMenuItem,
             this.discordToolStripMenuItem,
-            this.modSupportToolStripMenuItem,
             this.reportClientIssueToolStripMenuItem,
             this.reportMetadataIssueToolStripMenuItem,
             this.aboutToolStripMenuItem});
             this.helpToolStripMenuItem.Name = "helpToolStripMenuItem";
             this.helpToolStripMenuItem.Size = new System.Drawing.Size(61, 29);
             resources.ApplyResources(this.helpToolStripMenuItem, "helpToolStripMenuItem");
+            //
+            // gameDiscordToolStripMenuItem
+            //
+            this.gameDiscordToolStripMenuItem.Name = "gameDiscordToolStripMenuItem";
+            this.gameDiscordToolStripMenuItem.Size = new System.Drawing.Size(230, 30);
+            this.gameDiscordToolStripMenuItem.Click += new System.EventHandler(this.gameDiscordToolStripMenuItem_Click);
+            resources.ApplyResources(this.gameDiscordToolStripMenuItem, "gameDiscordToolStripMenuItem");
+            //
+            // modSupportToolStripMenuItem
+            //
+            this.modSupportToolStripMenuItem.Name = "modSupportToolStripMenuItem";
+            this.modSupportToolStripMenuItem.Size = new System.Drawing.Size(230, 30);
+            this.modSupportToolStripMenuItem.Click += new System.EventHandler(this.modSupportToolStripMenuItem_Click);
+            resources.ApplyResources(this.modSupportToolStripMenuItem, "modSupportToolStripMenuItem");
             //
             // userGuideToolStripMenuItem
             //
@@ -375,13 +393,6 @@ namespace CKAN.GUI
             this.discordToolStripMenuItem.Size = new System.Drawing.Size(230, 30);
             this.discordToolStripMenuItem.Click += new System.EventHandler(this.discordToolStripMenuItem_Click);
             resources.ApplyResources(this.discordToolStripMenuItem, "discordToolStripMenuItem");
-            //
-            // modSupportToolStripMenuItem
-            //
-            this.modSupportToolStripMenuItem.Name = "modSupportToolStripMenuItem";
-            this.modSupportToolStripMenuItem.Size = new System.Drawing.Size(230, 30);
-            this.modSupportToolStripMenuItem.Click += new System.EventHandler(this.modSupportToolStripMenuItem_Click);
-            resources.ApplyResources(this.modSupportToolStripMenuItem, "modSupportToolStripMenuItem");
             //
             // reportClientIssueToolStripMenuItem
             //
@@ -990,9 +1001,11 @@ namespace CKAN.GUI
         private System.Windows.Forms.ToolStripMenuItem GameCommandlineToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem compatibleGameVersionsToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem helpToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem gameDiscordToolStripMenuItem;
+        private System.Windows.Forms.ToolStripMenuItem modSupportToolStripMenuItem;
+        private System.Windows.Forms.ToolStripSeparator HelpToolStripSeparator;
         private System.Windows.Forms.ToolStripMenuItem userGuideToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem discordToolStripMenuItem;
-        private System.Windows.Forms.ToolStripMenuItem modSupportToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem reportClientIssueToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem reportMetadataIssueToolStripMenuItem;
         private System.Windows.Forms.ToolStripMenuItem aboutToolStripMenuItem;