+++ GUI/Model/ModList.cs
@@ -35,6 +35,7 @@ namespace CKAN.GUI
         /// <param name="allTags">All tag definitions</param>
         /// <param name="coreConfig">Core configuration</param>
         /// <param name="guiConfig">GUI configuration</param>
+        /// <param name="graphics">Graphics object for calculating word wrap</param>
         /// <param name="mc">Changes the user has made</param>
         /// <returns>The mod list</returns>
         public ModList(IReadOnlyCollection<GUIMod> modules,
@@ -43,12 +44,14 @@ namespace CKAN.GUI
                        ModuleTagList               allTags,
                        IConfiguration              coreConfig,
                        GUIConfiguration            guiConfig,
+                       Graphics                    graphics,
                        List<ModChange>?            mc = null)
         {
             this.allLabels        = allLabels;
             this.allTags          = allTags;
             this.coreConfig       = coreConfig;
             this.guiConfig        = guiConfig;
+            this.graphics         = graphics;
             activeSearches        = guiConfig.DefaultSearches
                                              ?.Select(s => ModSearch.Parse(allLabels, instance, s))
                                               .OfType<ModSearch>()
@@ -226,7 +229,16 @@ namespace CKAN.GUI
             var installSize   = new DataGridViewTextBoxCell { Value = mod.InstallSize             };
             var releaseDate   = new DataGridViewTextBoxCell { Value = mod.Module.release_date?.ToLocalTime() };
             var installDate   = new DataGridViewTextBoxCell { Value = mod.InstallDate             };
-            var desc          = new DataGridViewTextBoxCell { Value = ToGridText(mod.Abstract)    };
+            var desc          = new DataGridViewTextBoxCell
+                                {
+                                    Value       = ToGridText(mod.Abstract),
+                                    ToolTipText = mod.Description is { Length: > 0 }
+                                                      ? string.Join(Environment.NewLine,
+                                                                    graphics.WordWrap(mod.Description, 600)
+                                                                            .Prepend("")
+                                                                            .Prepend(mod.Abstract))
+                                                      : mod.Abstract,
+                                };
 
             item.Cells.AddRange(selecting, autoInstalled, updating, replacing, name, author, installVersion, latestVersion, compat, downloadSize, installSize, releaseDate, installDate, downloadCount, desc);
 
@@ -674,6 +686,7 @@ namespace CKAN.GUI
         private readonly ModuleTagList                  allTags;
         private readonly IConfiguration                 coreConfig;
         private readonly GUIConfiguration               guiConfig;
+        private readonly Graphics                       graphics;
         private          IReadOnlyCollection<ModSearch> activeSearches;
 
         private static readonly Color conflictColor = Color.FromArgb(255, 64, 64);